src/backend/lib/requirements.txt
uvicorn==0.23.2
requests==2.32.2
SQLAlchemy==1.4.48
mysql-connector-python==8.0.32
minio==7.2.10
weaviate==0.1.2
weaviate-client==3.26.7
python-multipart==0.0.17
httpx==0.28.1
fastapi==0.115.8
pydantic==2.10.6
clerk-backend-api==1.7.2
PyJWT==2.10.1

# Install CPU-specific PyTorch components
numpy<2
torch==2.0.1+cpu
torchvision==0.15.2+cpu
torchaudio==2.0.2+cpu
-f https://download.pytorch.org/whl/torch_stable.html

# Install sentence-transformers after CPU torch is available
sentence-transformers==3.4.1

